% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_vals.R
\name{convert_from_NA}
\alias{convert_from_NA}
\title{Convert values from NA to an alternative value}
\usage{
convert_from_NA(x, to, cols = NULL)
}
\arguments{
\item{x}{A data frame}

\item{to}{Values that NA values should be converted to}

\item{cols}{0 or more column names or positions.}
}
\description{
\code{convert_from_NA} finds NA values and changes them to a value specified
by \code{to}
}
\seealso{
\code{\link[sgcleanup]{convert_to_NA}}
}
