% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_vals.R
\name{convert_to_NA}
\alias{convert_to_NA}
\title{Convert values to NA}
\usage{
convert_to_NA(x, na_vals, cols = NULL)
}
\arguments{
\item{data}{A data frame}

\item{keys}{One or more values that should be converted to NA.}

\item{values}{0 or more column names or positions.}
}
\description{
\code{convert_to_NA} finds values sepcified by \code{na_vals} and sets them
to NA.
}
\seealso{
\code{\link[tidyr]{spread}}
}
