% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor.R
\name{refactor_by_idx}
\alias{refactor_by_idx}
\title{Create a factor based on specified level:index pairs}
\usage{
refactor_by_idx(x, levs, idx)
}
\arguments{
\item{x}{A vector that will be converted to a factor based with level order
specified by \code{levs}:\code{idx} pairing
\code{from}.}

\item{levs}{The levels contained in \code{x}}

\item{idx}{The order in which \code{levs} will be incorporated into the
new factor}
}
\value{
A factor with levels ordered according to \code{levs}:\code{idx}
pairing
}
\description{
\code{refactor_by_idx} changes \code{x} to a factor, and specifies level
order based the 1:1 pairing of \code{levs} and \code{idx}
}
