% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p_val.R
\name{round_tz}
\alias{round_tz}
\title{Round, leaving trailing zeroes}
\usage{
round_tz(x, digits = 2)
}
\arguments{
\item{x}{A numerical vector}

\item{digits}{digits to round to}
}
\value{
A vector of rounded numbers.
}
\description{
Round, leaving trailing zeroes
}
