% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr-extensions.R
\name{spread_sg}
\alias{spread_sg}
\title{Spread key-value pairs composed of multiple keys and values}
\usage{
spread_sg(data, keys, values)
}
\arguments{
\item{data}{A data frame}

\item{keys}{One or more column names.}

\item{values}{One or more column names.}
}
\description{
\code{spread_sg} allows the use of the \code{spread} function from
\code{tidyr} over multiple keys and values. Essentially, keys and values
will be combined into one key and value, then separated in the final data
frame as appropriate.
}
\seealso{
\code{\link[tidyr]{spread}}
}
