% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{sub_regex}
\alias{sub_regex}
\title{Substitute regular expressions within a vector with another string}
\usage{
sub_regex(x, from, to)
}
\arguments{
\item{x}{A character vector, that will be searched for a collection of
regular expressions (specified \code{to}) replaced by strings specified in
\code{from}.}

\item{from}{A vector of regexs potentially contained within \code{x} that are
to be substituted}

\item{to}{The corresponding strings that will be used to replace
elemnts of \code{from}.
converted to}
}
\value{
A character vector, derived from \code{x}, will have 0 or more
elements altered based on the specified \code{from}:\code{to} mapping
}
\description{
\code{sub_value} finds regular expressions within \code{x} specified by
\code{from} and substitutes them with strings specified in \code{to}. Regexs
specified by \code{from} will be replaced in a 1:1 or many:1 fashion. Only
regexs within \code{from} that are contained in \code{x} will be replaced.
}
