% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{sub_value}
\alias{sub_value}
\title{Substitute exact matches within a vector with another value}
\usage{
sub_value(x, from, to)
}
\arguments{
\item{x}{A vector of values, some (or all) of which are equivalent to values
in \code{from} and will be replaced by values in \code{to}}

\item{from}{A vector of values potentially contained within \code{x} that are
to be substituted}

\item{to}{The corresponding value (or values) that \code{from} will be
converted to}
}
\value{
A new vector, derived from \code{x}, with 0 or more values
substituted for new ones
}
\description{
\code{sub_value} finds exact matches within \code{x} specified by
\code{from} and substitutes them with values specified in \code{to}. Values
specified by \code{from} will be replaced in a 1:1 or many:1 fashion. Only
values within \code{from} that are contained in \code{x} will be replaced.
}
