% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation.R
\name{tabulate_wd}
\alias{tabulate_wd}
\title{Tabulate the columns of wide data}
\usage{
tabulate_wd(dat, cols)
}
\arguments{
\item{dat}{A data.frame}

\item{cols}{Columns to tabulate}
}
\value{
A data.frame with original \code{cols} as rows and columns containing
the counts of individual values
}
\description{
Counts unique values of specified columns and returns these counts as a
data.frame
}
