% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_vals.R
\name{convert_to_NA}
\alias{convert_to_NA}
\title{Convert values to NA}
\usage{
convert_to_NA(x, from, cols = NULL)
}
\arguments{
\item{x}{A data frame}

\item{from}{One or more values that should be converted to NA.}

\item{cols}{0 or more column names or positions.}
}
\description{
\code{convert_to_NA} finds values specified by \code{from} and sets them
to NA.
}
\seealso{
\code{\link[sgcleanup]{convert_from_NA}}
}
